requirements.txt
aiogram
requests
python-dotenv
apscheduler
feedparser